Clean up and adding Qodana to build.
