+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "5.0",
+    "version": "7.0.0",
     "rollForward": "latestMajor",
     "allowPrerelease": false
   }
