+++ qodana.yaml
@@ -1,0 +1,7 @@
+version: "1.0"
+linter: jetbrains/qodana-dotnet:2023.1-eap
+exclude:
+  - name: All
+    paths:
+      - shared/JetBrains.Annotations.cs
+      - test/Htmx.Sample
