+++ src/Htmx.TagHelpers/Htmx.TagHelpers.csproj
@@ -39,6 +39,12 @@
         <EmbeddedResource Include="JavaScript\antiforgerySnippet.js" LogicalName="AntiforgeryJavaScript" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Compile Include="..\..\shared\JetBrains.Annotations.cs">
+        <Link>JetBrains.Annotations.cs</Link>
+      </Compile>
+    </ItemGroup>
+
     <PropertyGroup>
         <MinVerSkip Condition="'$(Configuration)' == 'Debug'">true</MinVerSkip>
     </PropertyGroup>
