+++ src/Htmx.TagHelpers/HtmxSnippets.cs
@@ -1,23 +1,24 @@
 using System;
 using System.IO;
+using JetBrains.Annotations;
 
-namespace Htmx.TagHelpers
+namespace Htmx.TagHelpers;
+
+[PublicAPI]
+public static class HtmxSnippets
 {
-    public static class HtmxSnippets
+    public static string AntiforgeryJavaScript
+        => GetString(nameof(AntiforgeryJavaScript));
+
+    private static string GetString(string name)
     {
-        public static string AntiforgeryJavaScript
-            => GetString(nameof(AntiforgeryJavaScript));
+        var assembly = typeof(HtmxSnippets).Assembly;
+        using var resource = assembly.GetManifestResourceStream(name);
 
-        private static string GetString(string name)
-        {
-            var assembly = typeof(HtmxSnippets).Assembly;
-            using var resource = assembly.GetManifestResourceStream(name);
-
-            if (resource == null)
-                throw new ArgumentException($"Resource \"{name}\" was not found.", nameof(name));
+        if (resource == null)
+            throw new ArgumentException($"Resource \"{name}\" was not found.", nameof(name));
             
-            using var reader = new StreamReader(resource);
-            return reader.ReadToEnd();
-        }
+        using var reader = new StreamReader(resource);
+        return reader.ReadToEnd();
     }
 }
