+++ test/Htmx.Sample/Controllers/HomeController.cs
@@ -1,20 +1,19 @@
 using Microsoft.AspNetCore.Mvc;
 
-namespace Htmx.Sample.Controllers
-{
-    public class HomeController : Controller
-    {
-        [HttpGet, Route("home")]
-        public IActionResult Index(int id)
-        {
-            return Content($"<h2>Hello, From Home Controller (id: {id})</h2>");
-        }
+namespace Htmx.Sample.Controllers;
 
-        [ValidateAntiForgeryToken]
-        [HttpPost, Route("home/name", Name = "named")]
-        public IActionResult Named()
-        {
-            return Content("<h2>Hello, From Named Route</h2>");
-        }
+public class HomeController : Controller
+{
+    [HttpGet, Route("home")]
+    public IActionResult Index(int id)
+    {
+        return Content($"<h2>Hello, From Home Controller (id: {id})</h2>");
+    }
+
+    [ValidateAntiForgeryToken]
+    [HttpPost, Route("home/name", Name = "named")]
+    public IActionResult Named()
+    {
+        return Content("<h2>Hello, From Named Route</h2>");
     }
 }
