+++ test/Htmx.Sample/Pages/Index.cshtml.cs
@@ -6,24 +6,23 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.Extensions.Logging;
 
-namespace Htmx.Sample.Pages
+namespace Htmx.Sample.Pages;
+
+public class IndexModel : PageModel
 {
-    public class IndexModel : PageModel
+    private readonly ILogger<IndexModel> _logger;
+
+    public IndexModel(ILogger<IndexModel> logger)
     {
-        private readonly ILogger<IndexModel> _logger;
+        _logger = logger;
+    }
 
-        public IndexModel(ILogger<IndexModel> logger)
-        {
-            _logger = logger;
-        }
+    public void OnGet()
+    {
+    }
 
-        public void OnGet()
-        {
-        }
-
-        public IActionResult OnGetSnippet()
-        {
-            return Content("<h2>Hello, World!</h2>", "text/html");
-        }
+    public IActionResult OnGetSnippet()
+    {
+        return Content("<h2>Hello, World!</h2>", "text/html");
     }
 }
