+++ test/Htmx.Sample/Pages/_ViewImports.cshtml
@@ -1,4 +1,4 @@
 ﻿@using Htmx.Sample
-@namespace Htmx.Sample.Pages
+@namespace Htmx.Sample.Pages ;
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, Htmx.TagHelpers
